vertosphere/java: Add startup script of the daemon and rename main package.
